# Faster generic broadcasting