--- a/src/lib/broadcast.jl
+++ b/src/lib/broadcast.jl
@@ -168,9 +168,10 @@ collapse_nothings(xs::AbstractArray{Nothing}) = nothing
 collapse_nothings(xs) = xs
 
 _purefun(::Type{F}) where {F<:Function} = isempty(fieldnames(F))
-_purefun(::Type{ComposedFunction{F,G}}) where {F,G} = _purefun(F) && _purefun(G)
 _purefun(::Type) = false
-
+if VERSION >= v"1.6"
+  _purefun(::Type{ComposedFunction{F,G}}) where {F,G} = _purefun(F) && _purefun(G)
+end
 _purefun(::Type{typeof(^)}) = false  # fix @testset "power" & @testset "diagonal hessian"
 
 @adjoint function broadcasted(::AbstractArrayStyle, f::F, args...) where {F}