@ToucheSir 2021-06-20
With the caveat that I don't really understand what's going on, has this subsumed https://github.com/FluxML/Zygote.jl/pull/807?