@mcabbott 2021-06-20
I was unaware of #807. This does not try to make broadcasting better able to handle state; it does try to make it ignore state whenever possible, in the interest of not being slow. 

The last few commits add a similar test to `map`. Trying the test from the issue at the REPL, it fails to spot the state. Which seems not to happen in the tests, since they pass.
```
julia> s = 0;

julia> function f(x)
            global s += x
          end;

julia> gradient(x -> sum(map(f, x)), 1:10)
([1, 2, 3, 4, 5, 6, 7, 8, 9, 10],)  # this PR
([10, 9, 8, 7, 6, 5, 4, 3, 2, 1],)  # tagged version, test is == (10:-1:1,)

julia> fieldnames(typeof(f))
()

julia> Base.issingletontype(typeof(f))
true
```
Not sure what to make of that, global scope is weird? Maybe the example is too artificial, this is not actually a sensible gradient calculation.

The examples I invented (and added as tests here) do detect when a variable is closed over:
```
julia> gradient((x,y) -> sum((z->z^2+y[1]).(x)), [1,2,3], [4,5])
([2, 4, 6], [3, 0])

julia> fieldnames(typeof((z->z^2+y[1])))  # and Base.issingletontype is true
()

julia> let y = [1,2]
         fieldnames(typeof((z->z^2+y[1])))  # and Base.issingletontype is false
       end
(:y,)
```