@DhairyaLGandhi 2021-06-22
Pushing out StaticArrays would make DiffImages.jl harder to develop cc @somtambe