@DhairyaLGandhi 2021-06-26
How would this support immutable structures like static arrays without redoing work of writing kernels for every kind of array type that comes along?