@mcabbott 2021-06-26
Julia just specialised like usual. It's not mutating anything. What makes you think this is less likely to work on immutable arrays than the previous code?

There are right now zero tests for StaticArrays. But when I try things, it seems pretty easy to find examples where broadcasting fails, on the tagged version. Some get fixed by this change.