@oschulz 2021-10-01
>  28.625 μs (28 allocations: 469.50 KiB)    # this PR, reverse mode
>  19.375 μs (52 allocations: 314.17 KiB)    # this PR, forward mode

@mcabbott , how did you (meaning, how can I :-) ) select forward mode in Zygote - was this done using `Zygote.forwarddiff`?