@mcabbott 2021-10-01
The merged version will automatically use forward mode whenever possible. 

The times quoted were from different variants tried while writing this PR -- the first step was `map(StaticGetter{i}(), dxs_zip)` which improved the generic reverse mode, the second step was to call `broadcast_forward` when possible (this internal function was previously used only for CUDA stuff). 

Applying `Zygote.forwarddiff` to the whole broadcast will probably be much slower, as it works out the jacobian. While in broadcasting you know that each output only depends on certain inputs (the "sparsity pattern" is how some people like to say it), which `broadcast_forward` knows about.