@oschulz 2021-10-01
Thanks, @mcabbott !

> Applying Zygote.forwarddiff to the whole broadcast will probably be much slower

Oh, sure - I meant for the function broadcasted over.