@mcabbott 2021-10-01
Ah OK. I never thought to try that. In principle that ought to compile down to roughly the same thing, right? But in practice:
```
julia> struct Cbrt ignore end;

julia> (::Cbrt)(x) = Zygote.forwarddiff(cbrt, x)

julia> Base.issingletontype(typeof(Cbrt(pi)))  # hence Zygote will not use dual numbers
false

julia> g1 = @btime gradient(xs -> sum(broadcast(Cbrt(pi), xs)), $xs);
  2.143 ms (40035 allocations: 2.90 MiB)

julia> g2 = @btime gradient(xs -> sum(broadcast(cbrt, xs)), $xs);  # default, as above
  90.667 μs (13 allocations: 312.78 KiB)

julia> g1[1] ≈ g2[1]
true
```