@oschulz 2021-10-01
Oh, interesting! Zygote now outperforms `ForwardDiffPullbacks.fwddiff` (which I mainly built because Zygote used to be so slow with broadcasts, often) in this case:

```
julia> using Zygote, ForwardDiffPullbacks, BenchmarkTools
julia> xs = randn(10_000)

julia> @benchmark gradient(xs -> sum(broadcast(cbrt, xs)), $xs)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  100.410 μs …  1.835 ms  ┊ GC (min … max): 0.00% … 91.63%
 Time  (median):     105.558 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   118.242 μs ± 91.057 μs  ┊ GC (mean ± σ):  6.04% ±  7.26%

 Memory estimate: 312.78 KiB, allocs estimate: 13.

julia> @benchmark gradient(xs -> sum(broadcast(fwddiff(cbrt), xs)), $xs)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  150.934 μs …  1.446 ms  ┊ GC (min … max): 0.00% … 87.23%
 Time  (median):     155.647 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   166.397 μs ± 61.921 μs  ┊ GC (mean ± σ):  2.00% ±  4.95%

 Memory estimate: 156.48 KiB, allocs estimate: 10.
```

I guess `ForwardDiffPullbacks.fwddiff` may still speed things up if the broadcasted function contains calculations that have many steps, in cases where Zygote fails to use `Dual`s automatically (maybe not so often, anymore?). And `fwddiff` seems to reduce the GC overhead a bit.