@mcabbott 2021-10-01
I hadn't seen this package, thanks. Will have to understand what it does, to save memory. 

Xref also https://github.com/JuliaDiff/ChainRules.jl/issues/531, where there are some ideas on how to use `derivatives_given_output` to do better.

For now Zygote uses dual numbers a lot... well the list of conditions is fairly long. The arguments and output must be real numbers. You shouldn't be inside a second derivative. The function must not be a closure, for which `Base.issingletontype(typeof(f))` is its (imperfect) test.

> broadcasted function contains calculations that have many steps

Do you mean fusion, or just a complicated `f` in `f.(x)`?

Zygote remains always un-fused, i.e. `f.(g.(x))` is exactly like `y=g.(x); f.(y)` done one-by-one. But `(f∘g).(x)` should still use dual numbers.