@oschulz 2021-10-01
But in the [`ForwardDiffPullbacks.fwddiff` example](https://oschulz.github.io/ForwardDiffPullbacks.jl/stable/api/), `fwddiff` still yields a massive speedup:

```julia
julia> using Zygote, ForwardDiffPullbacks, StaticArrays, BenchmarkTools

julia> f = (xs...) -> (sum(map(x -> sum(map(x -> x^2, x)), xs)))

julia> xs = (2, (3, 4), SVector(5, 6, 7));
julia> Xs = map(x -> fill(x, 100), xs);

julia> @benchmark Zygote.gradient(($Xs...) -> sum(f.($Xs...)), $Xs...)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  403.472 μs …   5.890 ms  ┊ GC (min … max): 0.00% … 91.96%
 Time  (median):     452.634 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   474.041 μs ± 294.882 μs  ┊ GC (mean ± σ):  3.51% ±  5.18%

 Memory estimate: 143.06 KiB, allocs estimate: 3229.

julia> @benchmark Zygote.gradient(($Xs...) -> sum(fwddiff($f).($Xs...)), $Xs...)
BenchmarkTools.Trial: 10000 samples with 7 evaluations.
 Range (min … max):  4.543 μs … 393.863 μs  ┊ GC (min … max): 0.00% … 95.79%
 Time  (median):     5.010 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.637 μs ±  12.513 μs  ┊ GC (mean ± σ):  8.13% ±  3.62%

 Memory estimate: 9.11 KiB, allocs estimate: 64.
```

Whew, not all my work was in vain - for now .... ;-)