@oschulz 2021-10-01
>> broadcasted function contains calculations that have many steps
> Do you mean fusion, or just a complicated f in f.(x)?

Just a f that has many steps, loops, etc., stuff that will result in a long tape if Zygote reverse-mode diffs it.